third_party/blink/web_tests/http/tests/streams/transferable/reason.html

<!DOCTYPE html>
<meta charset="utf-8">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/helpers.js"></script>
<script>
'use strict';

// These tests verify the algorithm for passing through error reasons from one
// realm to another. We only test these for cancel on a ReadableStream, and
// assume that all the other places that should be using the same algorithm are
// doing so.

function getTransferredReason(originalReason) {
  return new Promise((resolve, reject) => {
    createTransferredReadableStream({
      cancel(reason) {
        resolve(reason);
      }
    }).then(rs => rs.cancel(originalReason))
      .catch(reject);
  });
}

for (const value of ['hi', '\t\r\n', 7, 3.0, undefined, null, true, false]) {
  promise_test(async () => {
    const reason = await getTransferredReason(value);
    assert_equals(reason, value, 'reason should match');
  }, `reason with a simple value of '${value}' should be preserved`);
}

for (const badType of [Symbol('hi'), _ => 'hi']) {
  promise_test(async () => {
    const reason = await getTransferredReason(badType);
    assert_equals(reason, undefined, 'reason should be undefined');
  }, `reason with a type of '${typeof badType}' should be squished to undefined`);
}

for (const badNumber of [NaN, Infinity]) {
  promise_test(async () => {
    const reason = await getTransferredReason(badNumber);
    assert_equals(reason, null, 'reason should be null');
  }, `number with a value of '${badNumber}' should be squished to null`);
}

promise_test(async () => {
  const reasonAsJson =
        `{"foo":[1,"col"],"bar":{"hoge":0.2,"baz":{},"shan":null}}`;
  const reason = await getTransferredReason(JSON.parse(reasonAsJson));
  assert_equals(JSON.stringify(reason), reasonAsJson,
                'object should be preserved');
}, 'objects that can be completely expressed in JSON should be preserved');

promise_test(async () => {
  const circularObject = {};
  circularObject.self = circularObject;
  const reason = await getTransferredReason(circularObject);
  assert_true(reason instanceof TypeError, 'a TypeError should be output');
}, 'objects that cannot be expressed in JSON should result in a TypeError');

promise_test(async () => {
  const originalReason = new TypeError('hi');
  const reason = await getTransferredReason(originalReason);
  assert_true(reason instanceof TypeError,
              'type should be preserved');
  assert_equals(reason.message, originalReason.message,
                'message should be preserved');
}, 'the type and message of a TypeError should be preserved');

promise_test(async () => {
  const originalReason = new TypeError('hi');
  originalReason.foo = 'bar';
  const reason = await getTransferredReason(originalReason);
  assert_false('foo' in reason,
               'foo should not be preserved');
}, 'other attributes of a TypeError should not be preserved');

promise_test(async () => {
  const originalReason = new TypeError();
  originalReason.message = {};
  const reason = await getTransferredReason(originalReason);
  assert_equals(reason.message, '', 'message should not be preserved');
}, 'a TypeError message should not be preserved if it is not a string');

promise_test(async () => {
  const originalReason = new TypeError();
  Object.defineProperty(originalReason, 'message', {
    get() { return 'words'; }
  });
  const reason = await getTransferredReason(originalReason);
  assert_equals(reason.message, '', 'message should not be preserved');
}, 'a TypeError message should not be preserved if it is a getter');

promise_test(async () => {
  const originalReason = new TypeError();
  delete originalReason.message;
  TypeError.prototype.message = 'inherited message';
  const reason = await getTransferredReason(originalReason);
  delete TypeError.prototype.message;
  assert_equals(reason.message, '', 'message should not be preserved');
}, 'a TypeError message should not be preserved if it is inherited');

promise_test(async () => {
  const originalReason = new DOMException('yes', 'AbortError');
  const reason = await getTransferredReason(originalReason);
  assert_true(reason instanceof DOMException,
              'reason should be a DOMException');
  assert_equals(reason.message, originalReason.message,
                'the messages should match');
  assert_equals(reason.name, originalReason.name,
                'the names should match');
}, 'DOMException errors should be preserved');

promise_test(async () => {
  const originalReason = new RangeError('nope');
  const reason = await getTransferredReason(originalReason);
  assert_equals(typeof reason, 'object', 'reason should have type object');
  assert_false(reason instanceof RangeError,
               'reason should not be a RangeError');
  assert_false(reason instanceof Error,
               'reason should not be an Error');
  assert_equals(reason.constructor, Object, 'reason should be an Object');
  assert_array_equals(Object.getOwnPropertyNames(reason), [],
                      'reason should have no properties');
}, 'RangeErrors should not be preserved');

</script>
